Add solution for "A. Bit++" Codeforces problem
Implements a solution for the "A. Bit++" problem from Codeforces Round 173 (Div. 2). The program evaluates a sequence of operations (++X, X++, --X, X--) to compute the final value of a variable based on the given input.

Includes utility macros and functions to streamline competitive programming workflow. The implementation ensures efficient handling of test cases and adheres to problem constraints.